+++ src/main/java/com/personal/tracker/controller/FXMLController.java
@@ -1,4 +1,4 @@
-package com.personal.tracker;
+package com.personal.tracker.controller;
 
 import java.net.URL;
 import java.util.ResourceBundle;
